requirements.txt
Kivy
argparse
RPi.GPIO
urllib3
opencv-python
imutils
numpy
future
arcgis
Pillow
pyzbar
gpiozero
vext
vext.gi
omxplayer-wrapper